refactor: project-wide transition to folder-based modules with lean mod.rs
