+++ aura-core/src/bitfield/logic.rs
@@ -84,6 +84,7 @@ impl Bitfield {
 }
 
 #[cfg(test)]
+#[cfg(test)]
 mod tests {
     use super::*;
