+++ aura-core/src/buffer_pool/logic.rs
@@ -41,6 +41,7 @@ impl BufferPool {
 }
 
 #[cfg(test)]
+#[cfg(test)]
 mod tests {
     use super::*;
