+++ aura-core/src/buffer_pool/mod.rs
@@ -1,3 +1,2 @@
 pub mod logic;
-pub mod udp;
 pub use logic::*;
