+++ aura-core/src/config/logic.rs
