+++ aura-core/src/magnet/logic.rs
@@ -68,6 +68,7 @@ impl Magnet {
 }
 
 #[cfg(test)]
+#[cfg(test)]
 mod tests {
     use super::*;
