+++ aura-core/src/nat/logic.rs
