+++ aura-core/src/net_util/logic.rs
