+++ aura-core/src/net_util/mod.rs
@@ -1,3 +1,2 @@
 pub mod logic;
-pub mod udp;
 pub use logic::*;
