+++ aura-core/src/peer_registry/logic.rs
@@ -77,6 +77,7 @@ impl Default for PeerRegistry {
 }
 
 #[cfg(test)]
+#[cfg(test)]
 mod tests {
     use super::*;
