+++ aura-core/src/piece_picker/logic.rs
@@ -136,6 +136,7 @@ impl PiecePicker {
 }
 
 #[cfg(test)]
+#[cfg(test)]
 mod tests {
     use super::*;
